fix: fix plurals and do type conversions
