+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-11-26
+
+### Added
+- Added tests for code compiler
+- Added more examples
+- Added type parsing corrections based on variable names and translate types from other languages like int, or float
+
+### Changed
+- Split `lex`, `parse`, and `compile` into different files
+- Moved compile function out of compile-arb-script
+- Limit exports in index file
+- Updated [README.md](README.md)
+
+### Security
+- Prevent code injections through typing variables
+
 ## [0.3.0] - 2025-11-24
 
 ### Added
