+++ package.json
@@ -10,7 +10,7 @@
     "url": "git+https://github.com/helpwave/internationlization.git"
   },
   "license": "MPL-2.0",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "type": "module",
   "files": [
     "dist"
